requirements.txt
tensorflow==2.5.0
pyttsx3==2.90
fastapi
uvicorn
gunicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy